requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
requests>=2.26.0
tqdm>=4.62.0
seaborn>=0.11.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# Para despliegue
flask>=2.0.0
gunicorn>=20.1.0
fastapi>=0.70.0
uvicorn>=0.15.0

# Para pruebas
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.9.0
flake8>=3.9.0
isort>=5.9.0

# Para mobile deployment
tensorflow>=2.8.0
onnx>=1.10.0
onnxruntime>=1.9.0

# Para visualización
plotly>=5.3.0
streamlit>=1.2.0

# Para docker
docker>=5.0.0

# Para cloud deployment
boto3>=1.18.0
google-cloud-storage>=1.42.0